requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
alembic>=1.14.0
httpx>=0.28.0
beautifulsoup4>=4.12
transmission-rpc>=7.0
libtorrent
yt-dlp
python-telegram-bot>=20.0
Pillow>=10.0
apscheduler>=3.11.0
